Add files via upload
This repo shows how to develop a simple app on Render. This is an extension of this repo [FcisArgyAnalysis](https://github.com/rhamiro94/MacroyFinanzas/blob/main/AnalisisFcisporcategoria.ipynb). Here we can know what Investment Fund Category performs better over time and inside each category the top 20 funds that are part of it.